microservices/orion-options-analyzer/requirements.txt
Flask
gunicorn
google-cloud-firestore
google-cloud-aiplatform
pandas
requests
finnhub-python
PyYAML
